requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.5.30
click==8.0.1
docopt==0.6.2
Flask==2.0.1
gunicorn==20.0.4
httpserver==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-telegram-bot==13.7
pytz==2021.1
six==1.16.0
tornado==6.1
tzlocal==3.0
Werkzeug==2.0.1
